requirements.txt
numpy==1.23.4
pandas==1.5.3
Pillow==9.3.0
pytest==7.2.0
python-dotenv==0.21.0
requests==2.27.1